requirements.txt
scapy-python3